// libs/enums/property.enum.ts
export enum PropertyCondition {
  NEW = 'NEW',
  USED = 'USED',
}

// ===== TYPE =====
export enum PropertyType {
  SEDAN = 'SEDAN',
  SUV = 'SUV',
  COUPE = 'COUPE',
  HATCHBACK = 'HATCHBACK',
  CONVERTIBLE = 'CONVERTIBLE',
  VAN = 'VAN',
  TRUCK = 'TRUCK',
  ELECTRIC = 'ELECTRIC',
  WAGON = 'WAGON',
  PICKUP = 'PICKUP',
  MINIVAN = 'MINIVAN',
  CROSSOVER = 'CROSSOVER',
}

// ===== STATUS =====
export enum PropertyStatus {
  ACTIVE = 'ACTIVE',           // Faol, ko'rinadi
  SOLD = 'SOLD',               // Sotilgan
  RENTED = 'RENTED',           // Hozir rent qilingan
  RESERVED = 'RESERVED',       // Band qilingan (to'lov kutilmoqda)
  MAINTENANCE = 'MAINTENANCE', // Ta'mirlashda
  INACTIVE = 'INACTIVE',       // Vaqtincha faol emas (agent tomonidan)
  DELETE = 'DELETE',           // O'chirilgan
}

// ===== BRAND =====
export enum PropertyBrand {
  TOYOTA = 'TOYOTA',
  BMW = 'BMW',
  MERCEDES = 'MERCEDES',
  AUDI = 'AUDI',
  HONDA = 'HONDA',
  FORD = 'FORD',
  CHEVROLET = 'CHEVROLET',
  NISSAN = 'NISSAN',
  HYUNDAI = 'HYUNDAI',
  KIA = 'KIA',
  VOLKSWAGEN = 'VOLKSWAGEN',
  TESLA = 'TESLA',
  LEXUS = 'LEXUS',
  MAZDA = 'MAZDA',
  SUBARU = 'SUBARU',
  JEEP = 'JEEP',
  RAM = 'RAM',
  GMC = 'GMC',
  PORSCHE = 'PORSCHE',
  LAND_ROVER = 'LAND_ROVER',
  VOLVO = 'VOLVO',
  INFINITI = 'INFINITI',
  ACURA = 'ACURA',
  CADILLAC = 'CADILLAC',
  BUICK = 'BUICK',
  CHRYSLER = 'CHRYSLER',
  DODGE = 'DODGE',
  MITSUBISHI = 'MITSUBISHI',
  GENESIS = 'GENESIS',
  ALFA_ROMEO = 'ALFA_ROMEO',
  MINI = 'MINI',
  JAGUAR = 'JAGUAR',
  MASERATI = 'MASERATI',
  BENTLEY = 'BENTLEY',
  ROLLS_ROYCE = 'ROLLS_ROYCE',
  FERRARI = 'FERRARI',
  LAMBORGHINI = 'LAMBORGHINI',
  ASTON_MARTIN = 'ASTON_MARTIN',
  MCLAREN = 'MCLAREN',
  BUGATTI = 'BUGATTI',
  OTHER = 'OTHER',
}

// ===== FUEL TYPE =====
export enum PropertyFuelType {
  GASOLINE = 'GASOLINE',
  DIESEL = 'DIESEL',
  ELECTRIC = 'ELECTRIC',
  HYBRID = 'HYBRID',
  PLUG_IN_HYBRID = 'PLUG_IN_HYBRID',
  HYDROGEN = 'HYDROGEN',
  ETHANOL = 'ETHANOL',
  BIODIESEL = 'BIODIESEL',
  COMPRESSED_NATURAL_GAS = 'COMPRESSED_NATURAL_GAS',
  LIQUEFIED_PETROLEUM_GAS = 'LIQUEFIED_PETROLEUM_GAS',
}